Reduce JDK-specific classes
Since Java 11, we can reduce the number of JDK specific classes and methods somewhat farther.
