+++ api/src/main/java16/org/jboss/marshalling/reflect/JDKSpecific.java
@@ -18,23 +18,10 @@
 
 package org.jboss.marshalling.reflect;
 
-import static java.lang.System.getSecurityManager;
-import static java.security.AccessController.doPrivileged;
-import static sun.reflect.ReflectionFactory.getReflectionFactory;
-
-import java.io.IOException;
-import java.io.ObjectInput;
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
-import java.io.ObjectStreamException;
 import java.lang.invoke.MethodHandle;
 import java.lang.invoke.MethodHandles;
 import java.lang.invoke.MethodType;
-import java.lang.reflect.Constructor;
 import java.lang.reflect.RecordComponent;
-import java.lang.reflect.UndeclaredThrowableException;
-import java.security.PrivilegedAction;
-import sun.reflect.ReflectionFactory;
 
 /**
  * JDK-specific classes which are replaced for different JDK major versions.  This one is for Java 16 only.
@@ -44,10 +31,6 @@ import sun.reflect.ReflectionFactory;
 final class JDKSpecific {
 
     private static final MethodHandles.Lookup LOOKUP = MethodHandles.lookup();
-
-    private static final ReflectionFactory reflectionFactory = getSecurityManager() == null ? getReflectionFactory() : doPrivileged(new PrivilegedAction<ReflectionFactory>() {
-        public ReflectionFactory run() { return ReflectionFactory.getReflectionFactory(); }
-    });
 
     /**
      * Return if this class is a record type.
@@ -108,122 +91,7 @@ final class JDKSpecific {
                     .asType(MethodType.methodType(Object.class, paramTypes));
             return constructorHandle.invokeWithArguments(args);
         } catch (Throwable e) {
-            throw new IllegalStateException("Error calling onstructor on record class " + recordType, e);
-        }
-    }
-
-    static Constructor<?> newConstructorForSerialization(Class<?> classToInstantiate, Constructor<?> constructorToCall) {
-        return reflectionFactory.newConstructorForSerialization(classToInstantiate, constructorToCall);
-    }
-
-    static final class SerMethods {
-        private final MethodHandle readObject;
-        private final MethodHandle readObjectNoData;
-        private final MethodHandle writeObject;
-        private final MethodHandle readResolve;
-        private final MethodHandle writeReplace;
-        private final Constructor<?> noArgConstructor;
-        private final Constructor<?> objectInputConstructor;
-
-        SerMethods(Class<?> clazz) {
-            readObject = reflectionFactory.readObjectForSerialization(clazz);
-            readObjectNoData = reflectionFactory.readObjectNoDataForSerialization(clazz);
-            writeObject = reflectionFactory.writeObjectForSerialization(clazz);
-            readResolve = reflectionFactory.readResolveForSerialization(clazz);
-            writeReplace = reflectionFactory.writeReplaceForSerialization(clazz);
-            Constructor<?> ctor;
-            Constructor<?> noArgConstructor = null;
-            try {
-                ctor = clazz.getDeclaredConstructor();
-                noArgConstructor = reflectionFactory.newConstructorForSerialization(clazz, ctor);
-            } catch (NoSuchMethodException ignored) {
-            }
-            this.noArgConstructor = noArgConstructor;
-            Constructor<?> objectInputConstructor = null;
-            try {
-                ctor = clazz.getDeclaredConstructor(ObjectInput.class);
-                objectInputConstructor = reflectionFactory.newConstructorForSerialization(clazz, ctor);
-            } catch (NoSuchMethodException ignored) {
-            }
-            this.objectInputConstructor = objectInputConstructor;
-        }
-
-        boolean hasWriteObject() {
-            return writeObject != null;
-        }
-
-        void callWriteObject(Object object, ObjectOutputStream outputStream) throws IOException {
-            try {
-                writeObject.invoke(object, outputStream);
-            } catch (IOException | RuntimeException | Error e) {
-                throw e;
-            } catch (Throwable e) {
-                throw new UndeclaredThrowableException(e);
-            }
-        }
-
-        boolean hasReadObject() {
-            return readObject != null;
-        }
-
-        void callReadObject(Object object, ObjectInputStream inputStream) throws IOException, ClassNotFoundException {
-            try {
-                readObject.invoke(object, inputStream);
-            } catch (IOException | ClassNotFoundException | RuntimeException | Error e) {
-                throw e;
-            } catch (Throwable e) {
-                throw new UndeclaredThrowableException(e);
-            }
-        }
-
-        boolean hasReadObjectNoData() {
-            return readObjectNoData != null;
-        }
-
-        void callReadObjectNoData(Object object) throws ObjectStreamException {
-            try {
-                readObjectNoData.invoke(object);
-            } catch (ObjectStreamException | RuntimeException | Error e) {
-                throw e;
-            } catch (Throwable e) {
-                throw new UndeclaredThrowableException(e);
-            }
-        }
-
-        boolean hasWriteReplace() {
-            return writeReplace != null;
-        }
-
-        Object callWriteReplace(Object object) throws ObjectStreamException {
-            try {
-                return writeReplace.invoke(object);
-            } catch (ObjectStreamException | RuntimeException | Error e) {
-                throw e;
-            } catch (Throwable e) {
-                throw new UndeclaredThrowableException(e);
-            }
-        }
-
-        boolean hasReadResolve() {
-            return readResolve != null;
-        }
-
-        Object callReadResolve(Object object) throws ObjectStreamException {
-            try {
-                return readResolve.invoke(object);
-            } catch (ObjectStreamException | RuntimeException | Error e) {
-                throw e;
-            } catch (Throwable e) {
-                throw new UndeclaredThrowableException(e);
-            }
-        }
-
-        Constructor<?> getNoArgConstructor() {
-            return noArgConstructor;
-        }
-
-        Constructor<?> getObjectInputConstructor() {
-            return objectInputConstructor;
+            throw new IllegalStateException("Error calling constructor on record class " + recordType, e);
         }
     }
 }
